✨ Add output stream to fix other output streams
To be more specific, this FilterOutputStream can be used to “fix” a
GzipOutputStream for continued use in Freenet.

Prior to Java 17, the “OS” field in the GZIP header was set to 0; Java 17
changed this to 0xff, to match the spec, which is lovely but breaks Freenet
(because it changes hashes calculated for keys).

By replacing a single byte in the GZIP-compressed output stream, everything can
be made as it was.